openshift: properly skip same image evaluation for ci builds

+++ kvirt/cluster/openshift/__init__.py
@@ -143,7 +143,8 @@ def same_release_images(version='stable', tag='4.12', pull_secret='openshift_pul
         existing = os.popen(f'{path}/openshift-install version').readlines()[2].split(" ")[2].strip()
     except:
         return False
-    if path != '.' and not existing.startswith('quay.io/openshift-release-dev/ocp-release'):
+    if os.path.abspath(path) != os.getcwd() and not existing.startswith('quay.io/openshift-release-dev/ocp-release')\
+       and not existing.startswith('registry.ci.openshift.org/ocp/release'):
         warning("Assuming your disconnected openshift-install has the correct version")
         return True
     offline = offline_image(version=version, tag=tag, pull_secret=pull_secret)
